Add test for long running iterations, and move cluster tests to test_cluster to match implementation.
